+++ tx-listener/block_cursor.go
@@ -6,10 +6,11 @@ import (
 	"math/big"
 	"sync"
 	"time"
-	"github.com/sirupsen/logrus"
+
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
+	log "github.com/sirupsen/logrus"
 )
 
 // TxListenerReceipt contains useful information about a receipt
@@ -71,7 +72,10 @@ func (e TxListenerErrors) Error() string {
 
 // ChainTracker keep track of block chain highest mined block
 type ChainTracker interface {
+	// Return ID of the chain being tracked
 	ChainID() *big.Int
+
+	// Return block number of the highest final block in the canonical chain
 	HighestBlock(ctx context.Context) (int64, error)
 }
 
@@ -97,7 +101,7 @@ func (t *BaseTracker) ChainID() *big.Int {
 	return big.NewInt(0).Set(t.chainID)
 }
 
-// HighestBlock returns highest mined block on the tracked chain
+// HighestBlock returns highest mined & considered filnal block on the tracked chain
 func (t *BaseTracker) HighestBlock(ctx context.Context) (int64, error) {
 	header, err := t.ec.HeaderByNumber(ctx, t.chainID, nil)
 	if err != nil {
@@ -106,7 +110,7 @@ func (t *BaseTracker) HighestBlock(ctx context.Context) (int64, error) {
 	if header.Number.Uint64() <= t.depth {
 		return 0, nil
 	}
-	return int64(header.Number.Uint64() - t.depth + 1), nil
+	return int64(header.Number.Uint64() - t.depth), nil
 }
 
 // Future is an element used to start a task and retrieve its result later
@@ -147,6 +151,10 @@ type BlockCursor struct {
 	// currentHead stands for the most advanced known mined block (we use it as cache so we do not need to always fetch Eth client for last mined block)
 	blockNumber, currentHead int64
 
+	// ticker and trigger are used to control the flow of fetch call for new mined blocks
+	ticker  *time.Ticker
+	trigger chan struct{}
+
 	// CurrentBlock on the cursor
 	blocks chan *TxListenerBlock
 	errors chan *TxListenerError
@@ -162,14 +170,16 @@ type BlockCursor struct {
 	conf Config
 }
 
-func newBlockCursorFromTracker(ec EthClient, t ChainTracker, blockNumber int64, conf Config) *BlockCursor {
+func newBlockCursorFromTracker(ec EthClient, t ChainTracker, startBlockNumber int64, conf Config) *BlockCursor {
 	return &BlockCursor{
 		ec:           ec,
 		t:            t,
-		blockNumber:  blockNumber,
+		blockNumber:  startBlockNumber,
 		currentHead:  0,
 		blocks:       make(chan *TxListenerBlock),
 		errors:       make(chan *TxListenerError),
+		ticker:       time.NewTicker(conf.BlockCursor.Backoff),
+		trigger:      make(chan struct{}, 1),
 		blockFutures: make(chan *Future, conf.BlockCursor.Limit),
 		closed:       make(chan struct{}),
 		closeOnce:    &sync.Once{},
@@ -214,10 +224,22 @@ func (bc *BlockCursor) Close() {
 	})
 }
 
+func (bc *BlockCursor) trig() {
+	select {
+	case bc.trigger <- struct{}{}:
+	default:
+		// already triggered
+	}
+}
+
 // feeder feeds the blockFutures channel
 // It manages the main cursor loop that fetch blocks & receipts from Eth client
 func (bc *BlockCursor) feeder() {
 	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// We trigger to start listener
+	bc.trig()
 feedingLoop:
 	for {
 		select {
@@ -225,22 +247,32 @@ feedingLoop:
 			// Cancel pending fetches and break loop
 			cancel()
 			break feedingLoop
-		default:
+		case <-bc.trigger:
 			if bc.blockNumber <= bc.currentHead {
-				// We are behind chain head meaning we have at leasdt one block to fetch
+				// We are behind chain head meaning we have at least one block to fetch
 				bc.blockFutures <- bc.fetchBlock(ctx, bc.blockNumber)
+
+				// Increment block position
 				bc.blockNumber++
+
+				// As a block was available we re-trigger
+				bc.trig()
 			} else {
-				// We are ahead of last known chain head, so we refresh it
+				// We are ahead of chain head so we refresh chain head
 				head, err := bc.t.HighestBlock(ctx)
-				logrus.Tracef("Highest block : %v", head)
+				log.WithFields(log.Fields{
+					"number": head,
+				}).Tracef("tx-listener: highest block")
 				if head > bc.currentHead {
-					// Chain has moved forward (meaning new blocks have been mined and are ready to be fetched)
+					// Chain has moved forward (meaning at least one new final block is ready to be fetched)
 					bc.currentHead = head
+
+					// We trigger
+					bc.trig()
 				} else {
 					// We are still ahead or something went wrong
 					if err != nil {
-						// If we got an error while retrieving chain head we notify it in a future
+						// If we got an error while retrieving chain highest final block we notify it in a future
 						bFuture := &Future{
 							res: make(chan interface{}),
 							err: make(chan error),
@@ -254,14 +286,16 @@ feedingLoop:
 
 						bc.blockFutures <- bFuture
 					}
-
-					// Chain has not move forward so we sleep before retrying (waiting for updates on the chain)
-					time.Sleep(bc.conf.BlockCursor.Backoff)
 				}
 			}
+		case <-bc.ticker.C:
+			// We trigger on every tick
+			bc.trig()
 		}
 	}
 	close(bc.blockFutures)
+	close(bc.trigger)
+	bc.ticker.Stop()
 }
 
 // fetchBlock fetch a block asynchronously
@@ -271,20 +305,28 @@ func (bc *BlockCursor) fetchBlock(ctx context.Context, blockNumber int64) *Futur
 		err: make(chan error),
 	}
 
-	logrus.Tracef("Trying to fetch block : %v", blockNumber)
+	log.WithFields(log.Fields{
+		"block.number": blockNumber,
+	}).Tracef("tx-listener: fetch block")
 	// Retrieve block in a separate goroutine
 	go func() {
 		defer bFuture.Close()
 
 		block, err := bc.ec.BlockByNumber(ctx, bc.ChainID(), big.NewInt(blockNumber))
 		if err != nil {
-			bFuture.err <- err
+			bFuture.err <- &TxListenerError{
+				bc.ChainID(),
+				err,
+			}
 			return
 		}
 
 		// Block should be available
 		if block == nil {
-			bFuture.err <- BlockMissingError(blockNumber)
+			bFuture.err <- &TxListenerError{
+				bc.ChainID(),
+				&BlockMissingError{blockNumber},
+			}
 			return
 		}
 
@@ -303,7 +345,10 @@ func (bc *BlockCursor) fetchBlock(ctx context.Context, blockNumber int64) *Futur
 		for i, rFuture := range rFutures {
 			select {
 			case err := <-rFuture.err:
-				bFuture.err <- err
+				bFuture.err <- &TxListenerError{
+					bc.ChainID(),
+					err,
+				}
 				return
 			case res := <-rFuture.res:
 				receipt := res.(*TxListenerReceipt)
@@ -332,12 +377,18 @@ func (bc *BlockCursor) fetchReceipt(ctx context.Context, txHash common.Hash) *Fu
 		defer future.Close()
 		receipt, err := bc.ec.TransactionReceipt(ctx, bc.ChainID(), txHash)
 		if err != nil {
-			future.err <- err
+			future.err <- &TxListenerError{
+				bc.ChainID(),
+				err,
+			}
 			return
 		}
 
 		if receipt == nil {
-			future.err <- ReceiptMissingError(txHash.Hex())
+			future.err <- &TxListenerError{
+				bc.ChainID(),
+				&ReceiptMissingError{txHash.Hex()},
+			}
 			return
 		}