+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+### Version 0.4.0
+
+*Unreleased*
+
+- [FEAT] Update `MultiEthClient` for dynamic Dial of Ethereum clients
+- [FEAT] Implement elements for *Dependy Injection* pattern for `MultiEthClient`
+- [FIX] Fix infinite recursivity error in `txlistener.BlockMissingError` and `txlistener.ReceiptMissingError`
+- [REFACTOR] Update `txlistener.BlockCursor` main loop to use `time.Ticker` and `trigger`
+- [TEST] Enhance tests for `txlistener.BlockCursor`
+- [FIX] Rollback fix from `v0.3.3`
+
 ### Version 0.3.3
 
 *Released on April 04th 2019*